err msg fix

+++ server/controllers/products.js
@@ -1,7 +1,7 @@
 const Products = require("../models/product");
 const fs = require("fs/promises");
 const StatusCodes = require("http-status-codes");
-const { BadRequest } = require("../errors/index");
+const { BadRequest, NotFoundError } = require("../errors/index");
 
 const getProducts = async (req, res) => {
   const data = await fs.readFile("./db/productStore.json", "utf8");
@@ -9,7 +9,7 @@ const getProducts = async (req, res) => {
   res.status(StatusCodes.OK).json({ success: true, parsedData });
 };
 
-const createProducts = async (req, res) => {
+const addProducts = async (req, res) => {
   const { id } = req.body;
 
   if (!id) {
@@ -22,6 +22,10 @@ const createProducts = async (req, res) => {
   const product = parsedData.find(
     (product) => product.productId === Number(id)
   );
+
+  if (!product) {
+    throw new NotFoundError(`no product with id: ${id}`);
+  }
 
   const addedToDB = await Products.create({ ...product });
 
@@ -41,7 +45,13 @@ const getCartProducts = async (req, res) => {
 const deleteProduct = async (req, res) => {
   const { id } = req.params;
   const { productName } = await Products.findOne({ productId: id });
+
+  if (!productName) {
+    throw new NotFoundError(`no product with id: ${id}`);
+  }
+
   const product = await Products.deleteOne({ productId: id });
+
   res.status(StatusCodes.OK).json({
     success: true,
     msg: `hello ${req.user.username.toUpperCase()} the product ${productName.toUpperCase()} has been removed from your cart `,
@@ -50,7 +60,7 @@ const deleteProduct = async (req, res) => {
 };
 
 module.exports = {
-  createProducts,
+  addProducts,
   getProducts,
   deleteProduct,
   getCartProducts,
